refactor: module name schedule -> tag
