+++ noweekend-core/core-domain/src/main/kotlin/noweekend/core/domain/tag/ScheduleReader.kt
@@ -1,4 +1,4 @@
-package noweekend.core.domain.schedule
+package noweekend.core.domain.tag
 
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
