+++ noweekend-core/core-domain/src/main/kotlin/noweekend/core/domain/tag/ScheduleRepository.kt
@@ -1,4 +1,4 @@
-package noweekend.core.domain.schedule
+package noweekend.core.domain.tag
 
 import java.time.LocalDateTime
