+++ noweekend-core/core-domain/src/main/kotlin/noweekend/core/domain/tag/Tag.kt
@@ -1,4 +1,4 @@
-package noweekend.core.domain.schedule
+package noweekend.core.domain.tag
 
 import noweekend.core.domain.util.IdGenerator
